Working on getting tests to pass
