+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "description": "A monorepo for all npm packages in the Tasit ecosystem",
   "devDependencies": {
-    "@nomiclabs/buidler": "^1.3.6",
+    "@nomiclabs/buidler": "^1.3.8",
     "@types/node": "^13.13.1",
     "@typescript-eslint/eslint-plugin": "^3.1.0",
     "@typescript-eslint/parser": "^3.1.0",
@@ -30,19 +30,20 @@
     "clean": "npx lerna run clean --stream && npx lerna clean --yes && rm -rf ./node_modules && npm run remove:package-lock",
     "publish": "npm run bootstrap && npm test && npx lerna publish",
     "bootstrap": "npx lerna bootstrap",
-    "buidler-evm:scoped:start": "npx lerna run buidler-evm:start --scope @tasit/contracts --stream",
-    "buidler-evm:scoped:stop": "npx lerna run buidler-evm:stop --scope @tasit/contracts --stream",
-    "buidler-evm:scoped:watch": "npx lerna run buidler-evm:watch --scope @tasit/contracts --stream",
-    "buidler-evm:start": "npx buidler node > buidler-evm.log &",
-    "buidler-evm:stop": "kill `ps ax|grep 'buidler node'|grep -v grep| awk '{print $1}'` 2> /dev/null; exit 0",
-    "buidler-evm:watch": "tail -f -n 100 buidler-evm.log",
-    "buidler:migrate": "npx lerna run compile:contracts --scope @tasit/contracts --stream",
-    "builder:migrate:rinkeby": "npx lerna run compile:rinkeby --scope @tasit/contracts --stream",
-    "prepare:blockchain": "npm run buidler-evm:stop && npm run buidler-evm:start && npm run buidler:migrate",
+    "blockchain:scoped:start": "npx lerna run blockchain:start --scope @tasit/contracts --stream",
+    "blockchain:scoped:stop": "npx lerna run blockchain:stop --scope @tasit/contracts --stream",
+    "blockchain:scoped:watch": "npx lerna run blockchain:watch --scope @tasit/contracts --stream",
+    "blockchain:start": "npx buidler node > buidler-evm.log &",
+    "blockchain:stop": "kill `ps ax|grep 'buidler node'|grep -v grep| awk '{print $1}'` 2> /dev/null; exit 0",
+    "blockchain:watch": "tail -f -n 100 buidler-evm.log",
+    "compile:contracts": "npx lerna run compile:contracts --scope @tasit/contracts --stream",
+    "compile:contracts:rinkeby": "npx lerna run compile:rinkeby --scope @tasit/contracts --stream",
+    "deploy:contracts": "npx lerna run deploy:contracts --scope @tasit/contracts --stream",
+    "prepare:blockchain": "npm run blockchain:stop && npm run blockchain:start && npm run compile:contracts && npm run deploy:contracts",
     "pretest": "npm run prepare:blockchain",
     "remove:package-lock": "npx lerna exec --stream --concurrency 1 -- rm -f package-lock.json",
     "test": "npx lerna run test --stream --concurrency 1",
-    "posttest": "npm run buidler-evm:stop",
+    "posttest": "npm run blockchain:stop",
     "start": "npm run clean:all && npm run bootstrap && npm run prepare:blockchain",
     "lint": "npx eslint './packages/*/src/{*.js,*.ts,**/*.js,**/*.ts,**/*.jsx,**/*.tsx}' --fix"
   },
