+++ packages/account/package.json
@@ -50,5 +50,5 @@
     "tslib": "^2.0.0",
     "typescript": "^3.9.5"
   },
-  "gitHead": "08af46b69dccad031c3230a563650278e8688006"
+  "gitHead": "b013a22bf9953710aa35664e9762237900b19e08"
 }
