+++ packages/action/package.json
@@ -56,5 +56,5 @@
     "tslib": "^2.0.0",
     "typescript": "^3.9.3"
   },
-  "gitHead": "08af46b69dccad031c3230a563650278e8688006"
+  "gitHead": "b013a22bf9953710aa35664e9762237900b19e08"
 }
