+++ packages/action/src/standards/index.ts
@@ -1,8 +1,9 @@
-import ERC20 from "./erc20"
-import ERC721 from "./erc721"
+import ERC20 from "./erc20";
+import ERC721 from "./erc721";
 
 const standards = {
-  ERC20, ERC721
-}
+  ERC20,
+  ERC721,
+};
 
 export default standards;
