+++ packages/contracts/buidler.config.5.js
@@ -17,7 +17,14 @@ task("accounts", "Prints the list of accounts", async () => {
 module.exports = {
   // This is a sample solc configuration that specifies which version of solc to use
   solc: {
-    version: "0.5.3",
+    // Default evm version should be greater than `byzantium`
+    // for the create2 opcode to be available
+    // even though Gnosis Safe contracts say ^0.5.3
+    // Could also have specified a non-default
+    // evm version while still using 0.5.3, but that
+    // seems like it could create more problems than it
+    // solves
+    version: "0.5.10",
   },
   paths: {
     sources: "./contracts/5",
