--- packages/contracts/scripts/sample-script.ts
@@ -1,29 +1,0 @@
-// We require the Buidler Runtime Environment explicitly here. This is optional 
-// but useful for running the script in a standalone fashion through `node <script>`.
-// When running the script with `buidler run <script>` you'll find the Buidler
-// Runtime Environment's members available in the global scope.
-import bre from "@nomiclabs/buidler";
-
-async function main() {
-  // Buidler always runs the compile task when running scripts through it. 
-  // If this runs in a standalone fashion you may want to call compile manually 
-  // to make sure everything is compiled
-  // await bre.run('compile');
-
-  // We get the contract to deploy
-  const Greeter = await ethers.getContractFactory("Greeter");
-  const greeter = await Greeter.deploy("Hello, Buidler!");
-
-  await greeter.deployed();
-
-  console.log("Greeter deployed to:", greeter.address);
-}
-
-// We recommend this pattern to be able to use async/await everywhere
-// and properly handle errors.
-main()
-  .then(() => process.exit(0))
-  .catch(error => {
-    console.error(error);
-    process.exit(1);
-  });
