+++ packages/contracts/test/sample-test.js
@@ -1,3 +1,5 @@
+// Note: Using chai here with mocha rather
+// than jest
 const { expect } = require("chai");
 
 describe("Greeter", function() {
@@ -6,9 +8,9 @@ describe("Greeter", function() {
     const greeter = await Greeter.deploy("Hello, world!");
     
     await greeter.deployed();
-    expect(await greeter.greet()).toEqual("Hello, world!");
+    expect(await greeter.greet()).to.equal("Hello, world!");
 
     await greeter.setGreeting("Hola, mundo!");
-    expect(await greeter.greet()).toEqual("Hola, mundo!");
+    expect(await greeter.greet()).to.equal("Hola, mundo!");
   });
 });
