+++ packages/test-helpers/package.json
@@ -40,5 +40,6 @@
   "dependencies": {
     "@tasit/action": "^0.0.29",
     "ethers": "^4.0.47"
-  }
+  },
+  "gitHead": "b013a22bf9953710aa35664e9762237900b19e08"
 }
